requirements.txt
django
requests
prettyconf
pillow
easy-thumbnails
WeasyPrint
django-rosetta
djangorestframework